safari-600.7-branch/LayoutTests/fast/dom/Geolocation/timestamp-expected.txt

Tests that Geoposition timestamps are well-formed (non-zero and in the same units as Date.getTime).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS now != 0 is true
PASS t != 0 is true
PASS then != 0 is true
PASS now - 1 <= t is true
PASS t <= then + 1 is true
PASS successfullyParsed is true

TEST COMPLETE
